# Optimize remote scatter kernel + GPU-driven Synchronous execution