--- a/p2p/include/common.hpp
+++ b/p2p/include/common.hpp
@@ -40,7 +40,7 @@
 #define ASSUME_WR_IN_ORDER
 #define NUMA_AWARE_SCHEDULING
 #define ENABLE_PROXY_CUDA_MEMCPY
-// #define SYNCHRONOUS_COMPLETION
+#define SYNCHRONOUS_COMPLETION
 #define kQueueSize 1024
 #define kQueueMask (kQueueSize - 1)
 #define kMaxInflight 32