--- a/p2p/include/copy_ring.hpp
+++ b/p2p/include/copy_ring.hpp
@@ -30,38 +30,27 @@ struct CopyRing {
 
   bool emplace(CopyTask const& t) {
     uint32_t h = head.load(std::memory_order_relaxed);
-    uint32_t n = h + 1;
+    uint32_t n = (h + 1) & (COPY_RING_CAP - 1);
     if (n == tail.load(std::memory_order_acquire)) return false;  // full
-    buf[h & (COPY_RING_CAP - 1)] = t;
+    buf[h] = t;
     head.store(n, std::memory_order_release);
     emplace_count.fetch_add(1, std::memory_order_relaxed);
     return true;
   }
 
-  bool emplace(std::vector<CopyTask> const& tasks, uint32_t count) {
+  bool emplace(std::vector<CopyTask> const& tasks) {
     if (tasks.empty()) return true;
-    if (tasks.size() != count) {
-      fprintf(stderr, "Error: tasks.size() %zu does not match count %u\n",
-              tasks.size(), count);
-      std::abort();
-    }
 
     uint32_t h = head.load(std::memory_order_relaxed);
     uint32_t t = tail.load(std::memory_order_acquire);
-    uint32_t used = (h - t) & (COPY_RING_CAP - 1);
-    uint32_t free_slots = COPY_RING_CAP - used - 1;
-    if (count > free_slots) return false;
+    uint32_t cap = COPY_RING_CAP;
+    uint32_t free_slots = (t + cap - h - 1) & (cap - 1);
+    if (tasks.size() > free_slots) return false;
 
     uint32_t idx = h;
     for (CopyTask const& task : tasks) {
-      if (task.dst_dev < 0 || task.dst_dev >= NUM_GPUS) {
-        fprintf(stderr,
-                "Error: emplace task.dst_dev %d is out of range (0-%d)\n",
-                task.dst_dev, NUM_GPUS - 1);
-        std::abort();
-      }
-      buf[idx & (COPY_RING_CAP - 1)] = task;
-      idx++;
+      buf[idx] = task;
+      idx = (idx + 1) & (cap - 1);
     }
 
     head.store(idx, std::memory_order_release);
@@ -72,9 +61,9 @@ struct CopyRing {
 
   CopyTask* pop() {
     uint32_t t = tail.load(std::memory_order_relaxed);
-    if (t == head.load(std::memory_order_acquire)) return nullptr;
-    CopyTask* ret = &buf[t & (COPY_RING_CAP - 1)];
-    tail.store(t + 1, std::memory_order_release);
+    if (t == head.load(std::memory_order_acquire)) return nullptr;  // empty
+    CopyTask* ret = &buf[t];
+    tail.store((t + 1) & (COPY_RING_CAP - 1), std::memory_order_release);
     pop_count.fetch_add(1, std::memory_order_relaxed);
     return ret;
   }
@@ -90,26 +79,12 @@ struct CopyRing {
     size_t count = std::min(n, available);
 
     for (size_t i = 0; i < count; ++i) {
-      tasks.push_back(buf[t & (COPY_RING_CAP - 1)]);
-      t = t + 1;
-
-      if (tasks.back().dst_dev < 0 || tasks.back().dst_dev >= NUM_GPUS) {
-        fprintf(stderr, "Error: popN task.dst_dev %d is out of range (0-%d)\n",
-                tasks.back().dst_dev, NUM_GPUS - 1);
-        std::abort();
-      }
+      tasks.push_back(buf[t]);
+      t = (t + 1) & (COPY_RING_CAP - 1);  // Wraparound
     }
     pop_count.fetch_add(static_cast<uint32_t>(count),
                         std::memory_order_relaxed);
     tail.store(t, std::memory_order_release);
     return count;
   }
-
-  bool has_completed(uint32_t idx) {
-    return tail.load(std::memory_order_acquire) >= idx;
-  }
-
-  uint32_t get_head() const { return head.load(std::memory_order_acquire); }
-
-  uint32_t get_tail() const { return tail.load(std::memory_order_acquire); }
 };