--- a/p2p/include/peer_copy.cuh
+++ b/p2p/include/peer_copy.cuh
@@ -4,6 +4,11 @@
 #include "copy_ring.hpp"
 #include <cuda_runtime.h>
 
+template <typename X, typename Y, typename Z = decltype(X() + Y())>
+__host__ __device__ constexpr Z divUp(X x, Y y) {
+  return (x + y - 1) / y;
+}
+
 cudaError_t launch_peer_bulk_copy(void* dst_ptr, int dst_dev, void* src_ptr,
                                   int src_dev, size_t bytes,
                                   cudaStream_t stream = 0);
@@ -15,3 +20,8 @@ cudaError_t launch_peer_bulk_copy2(CopyTask const* host_tasks, int num_tasks,
 __global__ void peer_copy_kernel_vec_batched(CopyTask const* __restrict__ tasks,
                                              int num_tasks,
                                              int tasks_per_block);
+
+template <int PIPE_DEPTH,  // same as kPipelineDepth
+          typename VecT>   // 16 B per transaction
+__global__ void peer_copy_kernel_vec_pipelined(
+    CopyTask const* __restrict__ tasks, int num_tasks, int tasks_per_block);