Synchronous (wait for cmd to complete. Latency includes local post cmds to cpu proxy, gpu-direct rdma, nvlink forwarding, notification sent back to local cpu proxy.)
```
Overall avg GPU-measured latency  : 67.817 µs
Total cycles                      : 531734350109
Total ops                         : 3959978
End-to-end wall-clock time        : 1837.928 ms
Ops Throughput                    : 2.15 Mops
Total Throughput                  : 141.20 Gbps
```
4 SMs per GPU for NVLink forwarding. 

Asynchronous (proceed without waiting for cmd to complete, latency only measures cmd posting time + when gpu-direct rdma finishes, without waiting for notification to be heard)
```
Overall avg GPU-measured latency  : 28.320 µs
Total cycles                      : 222048253381
Total ops                         : 3959967
End-to-end wall-clock time        : 764.640 ms
Ops Throughput                    : 5.18 Mops
Total Throughput                  : 339.40 Gbps
```
8 SMs per GPU for NVLink forwarding.